Fix warnings.
GitOrigin-RevId: dbfa557bcf9dc3cd63dd5cbbbed9d23b48a5b627

+++ tdutils/td/utils/port/MemoryMapping.cpp
@@ -84,7 +84,7 @@ Result<MemoryMapping> MemoryMapping::create_from_file(const FileFd &file_fd, con
   auto data_offset = begin - fixed_begin;
   auto data_size = end - fixed_begin;
 
-  void *data = mmap(nullptr, data_size, PROT_READ, MAP_PRIVATE, fd, fixed_begin);
+  void *data = mmap(nullptr, narrow_cast<size_t>(data_size), PROT_READ, MAP_PRIVATE, fd, narrow_cast<off_t>(fixed_begin));
   if (data == MAP_FAILED) {
     return OS_ERROR("mmap call failed");
   }
